+++ src/core/log/middlewares/filter/interface.ts
@@ -1,0 +1,13 @@
+/*!
+ * V4Fire Core
+ * https://github.com/V4Fire/Core
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Core/blob/master/LICENSE
+ */
+
+import type { LogEvent } from 'core/log/middlewares';
+
+export interface LogFilter {
+	check(event: LogEvent): boolean;
+}